plate_historic test now refactoring and tested
