+++ tests/test_plate_historic_db.py
@@ -4,6 +4,7 @@ from BioPlate.database.plate_db import PlateDB
 from BioPlate.plate import Plate
 from pathlib import Path, PurePath
 import contextlib
+import numpy as np
 
 
 
@@ -48,7 +49,7 @@ class TestPlateDB(unittest.TestCase):
         This function is run every time at the beginning of each test
         :return:
         """
-        self.plate_list = self.phi.get_one_hplate(96)
+        self.plate_list = self.phi.get_hplate(numWell=96)
         self.plate = self.plate_list[0]
 
     def tearDown(self):
@@ -60,52 +61,30 @@ class TestPlateDB(unittest.TestCase):
 
     def test_get_hplate(self):
         self.assertIsInstance(self.plate_list, list, "plate_db.get_plate don't return a list")
-        self.assertEquals('<plate N°1 : 96-12-8>', str(self.plate),
+        self.assertEqual(f'<plate N°1: First plate to test, 96 wells, {self.phi.date_now}>', str(self.plate),
                           "plate_db.get_plate don't return the appropriate format")
-        self.assertTrue(str(type(self.plate)) == "<class 'BioPlate.database.plate_db.PlateDB.PlateDatabase'>",
+        self.assertTrue(str(type(self.plate)) == "<class 'BioPlate.database.plate_historic_db.PlateHist.PlateHistoric'>",
                         f"plate_db.get_plate don't return the right class : {str(type(self.plate))}")
 
     def test_plate_class(self):
-        self.assertEqual(self.plate.numWell, 96, "Error numWell association fail")
-        self.assertEqual(self.plate.numColumns, 12, "Error numColumns association fail")
-        self.assertEqual(self.plate.numRows, 8, "Error numRows association fail")
-        self.assertEqual(self.plate.surfWell, 0.29, "Error numRows association fail")
-        self.assertEqual(self.plate.maxVolWell, 200, "Error numRows association fail")
-        self.assertEqual(self.plate.workVolWell, 200, "Error numRows association fail")
-        self.assertEqual(self.plate.refURL, "https://csmedia2.corning.com/LifeSciences/Media/pdf/cc_surface_areas.pdf",
-                         "Error refURL association fail")
+    	self.assertEqual(self.plate.numWell, 96, "Error numWell association fail")
+    	self.assertEqual(self.plate.plate_name, "First plate to test", "Error numColumns association fail")
+    	np.testing.assert_array_equal( self.plate.plate_array, self.plt.plate)
+    	
+    def test_add_hplate(self):
+    	add_plate_1 = self.pdb.add_plate(numWell=6, numColumns=3, numRows=2, surfWell=9.5,  maxVolWell=2000, workVolWell=2000, refURL= 'https://csmedia2.corning.com/LifeSciences/Media/pdf/cc_surface_areas.pdf')
+    	p6 = self.pdb.get_one_plate(6)
+    	Plate6 = Plate(6, db_name="test_plate.db")
+    	add_hplate_1 = self.phi.add_hplate(Plate_id=Plate6.plates.id, numWell=Plate6.plates.numWell, plate_name="second plate", plate_array=Plate6.plate) 
+    	add_hplate_2 = self.phi.add_hplate(Plate_id=Plate6.plates.id, numWell=Plate6.plates.numWell, plate_name="second plate", plate_array=Plate6.plate)
+    	self.assertEqual(add_hplate_1, "plate second plate with 6  added to database plate historic")
+    	self.assertEqual(add_hplate_2, "plate already exist : 2")
+    	self.assertEqual(f'<plate N°2: second plate, 6 wells, {self.phi.date_now}>', str(self.phi.get_one_hplate(6)))
+    	self.assertEqual(6, self.phi.get_hplate(numWell=6)[0].numWell)
 
-    def test_add_plate(self):
-        add_plate_1 = self.pdb.add_plate(numWell=6,
-                      numColumns=3,
-                      numRows=2,
-                      surfWell=9.5,
-                      maxVolWell=2000,
-                      workVolWell=2000,
-                      refURL='https://csmedia2.corning.com/LifeSciences/Media/pdf/cc_surface_areas.pdf')
-        self.assertEqual(add_plate_1,
-        	                "plate with 6 added to the database")
-        add_plate_2 = self.pdb.add_plate(numWell=6,
-                      numColumns=3,
-                      numRows=2,
-                      surfWell=9.5,
-                      maxVolWell=2000,
-                      workVolWell=2000,
-                      refURL='https://csmedia2.corning.com/LifeSciences/Media/pdf/cc_surface_areas.pdf')
-        self.assertEqual(add_plate_2, None)
-        self.assertEquals('<plate N°2 : 6-3-2>', str(self.pdb.get_plate(6)[0]))
-        self.assertEquals(6, self.pdb.get_plate(6)[0].numWell)
-
-    def test_delete_plate(self):
-        self.pdb.add_plate(numWell=24,
-              numColumns=6,
-              numRows=4,
-              surfWell=0.33,
-              maxVolWell=400,
-              workVolWell=400,
-              refURL='https://csmedia2.corning.com/LifeSciences/Media/pdf/cc_surface_areas.pdf')
-        self.assertEqual(self.pdb.delete_plate(24), "plate with 24 numWell deleted")
-        self.assertEqual(self.pdb.get_plate(24), [])
+    def test_delete_hplate(self):        
+        self.assertEqual(self.phi.delete_hplate(6), "plate with 6 numWell deleted")
+        self.assertEqual(self.phi.get_hplate( numWell=6), [])
 
 if __name__ == '__main__':
     unittest.main()
